base-files: save default sysctl on start
In order to get the default value of a sysctl, a reboot must be carried
out beforehand or it must be looked up in the linux source code what the
default value is. To simplify this, all values of /proc/sys are saved
with the sysctl command into a temp file before any changes are by the
sysctl command.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>