build: propagate errors in scan and feed update
Currently, when an error occurs during scanning of a package Makefile, an
error message is logged, but the error is not actually propagated. This
makes it easy for build scripts to accidentally proceed based on an incomplete
package scan.

Propagate errors correctly in scan.mk and update_index in the feeds script.

Signed-off-by: Karsten Sperling <ksperling@apple.com>