ipq806x: Ethernet interface fixes for AP3935
Some fixes to Ethernet interfaces on top of #13030.

* Fix LAN1 fixed-link issue by wiring up pins in DTS, setting rgmii-id
mode. Without this, rgmii mode alone doesn't work bidirectionally.
With this commit applied, the LAN1 interface autonegotiates properly
and works great.

* Includes a quilt patch against the dwmac driver to "add" rgmii-id
mode support. I say "add" as it only adds support for that enum value
in certain validation modes -- the driver as-is won't initialize if
the DTS sets the "rgmii-id" mode, but the MAC layer is really agnostic
to which rgmii submode is used. This makes it accept any of the
valid submodes. I will attempt to upstream this patch.

* Sets the LAN2 MAC address in the DTS by deriving it from LAN1's
address. This isn't *exactly* how the factory firmware works, but
in all of the device samples I've checked (~10) it derives the
correct MAC without requiring nvmem_u-boot-env to support parsing
MAC addresses from multiple fields.

* Cleans up mdio0 in the DTS so it uses existing defined mdio.

* Setting both Ethernet interface MACs from the DTS is useful because
they will then consistently enumerate and LAN1 will always be eth0,
LAN2 will always be eth1. Also updates 02_network to ensure that
interfaces are tagged correctly (and consistently with LED config
from the prior patch).

Signed-off-by: Bryan Berg <bdb@north-eastham.org>