dropbear: use ssh-keygen  as an alias for dropbearkey
The DropBear's dropbearkey supports limited set of arguments of OpenSSH ssh-keygen:  -t, -q -N -Y
After the change you can generate a key with the same command.
Still many features of the original OpenSSH ssh-keygen are absent in the dropbearkey.
If it's needed then users should install openssh-keygen package that will replace the /usr/bin/ssh-keygen with the full version.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>