scripts/kernel_upgrade: fix workaround for merge/restore commit creation
Return part of the original design.  Issue is at some future time,
it could be possible for a versioned file to be created inside a
versioned directory.  Presently there are no cases of
"dir-<version>/file-<version>", but it seems best to be prepared for
this.

Alas, this points out `git fast-import` isn't really a proper binary
protocol.  There really needs to be a -0 option to allow arbitrary
filenames.

Signed-off-by: Elliott Mitchell <ehem+openwrt@m5p.com>