tools: pkgconf: provide prefix for host and hostpkg stagind directory
Some package might require to fix their pkg-config file to point to host
or hostpkg file. This is the case for glib2 library that provides with
pkg-config variables, tools to generates files from xml. Those tools
should use the host binary instead of the targets one to correctly build
packages that makes use of such tools.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>