rampis: mt7621: Disable toggling of kernel slots
Uboot selects which kernel slot to boot based on a flag in the factory
mtd partition. This is done in ubnt.sh which toggles the flags on each
sysupgrade.

With a single kernel layout this flag should always be set to kernel.
Also ensure that kernel is always flashed to the 6MB kernel1 slot.

Signed-off-by: Tim Lunn <tim@feathertop.org>